Premier test mais il y a des erreurs

+++ app/src/main/java/com/llab/ligablo/database/dao/etabs/AdresseDao.java
@@ -18,7 +18,7 @@ public interface AdresseDao {
     int getLastAdresse();
 
     @Insert
-    int insertAdresse(Adresse adresse);
+    long insertAdresse(Adresse adresse);
 
     @Update
     int updateAdresse(Adresse adresse);
